Stripe appearance improvements (#3795)
* Improve blocks checkout appearance

- Send background color to Stripe
- Fix text color
- Introduce style variables

* Add shortcode checkout selectors

* Use correct checkout type when computing styles

* Switch text color and theme based on background color

* Remove text customizations

We already have a global text color in variables. Use of label styles
for Text doesn't quite work for dark themes. So we remove the
customization here so that it uses default Stripe styles (gray text)
which looks good on both light/dark themes.

* Reintroduce Text styling

This time using paragraph rules instead of label rules

* Change text color based on background color

* Add changelog entries

* Fix tests